Sal branch afternoon

+++ src/main/java/com/foodocs/service/MenuDAO.java
@@ -7,8 +7,8 @@ import java.util.List;
 
 public interface MenuDAO {
     public Menu getMenu();
-    public MenuItem getMenuItem(long id);
-    public MenuItem getIngredients(long id);
+    public MenuItem getMenuItem(int id);
+    public MenuItem getIngredients(int id);
     public void add(MenuItem menuItem);
     public void remove(MenuItem menuItem);
     public void update(MenuItem menuItem);
